backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==4.2.23
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
environ==1.0
idna==3.10
pdf2image==1.17.0
pillow==11.3.0
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
